refactor(sdk/go): nest package under sdk/go/openmeter
Move the package to sdk/go/openmeter so the import base matches the package
name (github.com/openmeterio/openmeter/sdk/go/openmeter), letting callers use
openmeter.New without a goimports-inserted alias. Module root stays sdk/go.
