+++ sdk/go/openmeter/query_test.go
